refactor(bimi): 🔨 split parsing into partial class
