+++ DomainDetective/Protocols/BimiAnalysis.cs
@@ -22,7 +22,7 @@ namespace DomainDetective {
     /// that reference an SVG logo file and an optional certificate. This class
     /// validates presence and accessibility of those resources.
     /// </remarks>
-public class BimiAnalysis {
+public partial class BimiAnalysis {
         /// <summary>Gets the concatenated BIMI record text.</summary>
         public string? BimiRecord { get; private set; }
         /// <summary>Gets a value indicating whether a BIMI record was found.</summary>
@@ -76,26 +76,7 @@ public class BimiAnalysis {
         public async Task AnalyzeBimiRecords(IEnumerable<DnsAnswer> dnsResults, InternalLogger logger, CancellationToken cancellationToken = default) {
             await Task.Yield();
 
-            BimiRecord = null;
-            BimiRecordExists = false;
-            StartsCorrectly = false;
-            Location = null;
-            Authority = null;
-            LocationUsesHttps = false;
-            AuthorityUsesHttps = false;
-            DeclinedToPublish = false;
-            InvalidLocation = false;
-            SvgFetched = false;
-            SvgValid = false;
-            SvgInvalidReason = null;
-            SvgSizeValid = false;
-            DimensionsValid = false;
-            ViewBoxValid = false;
-            ValidVmc = false;
-            VmcSignedByKnownRoot = false;
-            VmcContainsLogo = false;
-            VmcCertificate = null;
-            FailureReason = null;
+            ResetState();
 
             if (dnsResults == null) {
                 logger?.WriteVerbose("DNS query returned no results.");
@@ -112,37 +93,7 @@ public class BimiAnalysis {
             BimiRecord = string.Join(" ", recordList.Select(r => r.Data));
             logger.WriteVerbose($"Analyzing BIMI record {BimiRecord}");
 
-            StartsCorrectly = BimiRecord?.StartsWith("v=BIMI1", StringComparison.OrdinalIgnoreCase) == true;
-
-            foreach (var part in (BimiRecord ?? string.Empty).Split(';')) {
-                var kv = part.Split(new[] { '=' }, 2);
-                if (kv.Length != 2) {
-                    continue;
-                }
-
-                var key = kv[0].Trim();
-                var value = kv[1].Trim();
-
-                switch (key) {
-                    case "l":
-                        Location = value;
-                        InvalidLocation = !(value.StartsWith("https://", StringComparison.OrdinalIgnoreCase)
-                            && (value.EndsWith(".svg", StringComparison.OrdinalIgnoreCase)
-                                || value.EndsWith(".svgz", StringComparison.OrdinalIgnoreCase)));
-                        if (InvalidLocation) {
-                            logger?.WriteWarning("Invalid BIMI indicator location {0}", value);
-                        }
-                        break;
-                    case "a":
-                        Authority = value;
-                        break;
-                }
-            }
-
-            LocationUsesHttps = string.IsNullOrEmpty(Location) || Location.StartsWith("https://", StringComparison.OrdinalIgnoreCase);
-            AuthorityUsesHttps = string.IsNullOrEmpty(Authority) || Authority.StartsWith("https://", StringComparison.OrdinalIgnoreCase);
-
-            DeclinedToPublish = string.IsNullOrEmpty(Location) && string.IsNullOrEmpty(Authority);
+            ParseBimiHeader(BimiRecord!, logger);
 
             if (!string.IsNullOrEmpty(Location) && !InvalidLocation) {
                 if (!LocationUsesHttps) {
